SBERDOMA-1122 moved addressMeta field GraphQL declaration into "property" domain
